Added models
Models to be used in the ParameterOptimization.py script. 
Model M019 is the default model described in the Methods of the main text. 
Model M021 is the default model (M019) with a Michaelis-Menten dependency of downstream target production on p53-p, i.e., the hill parameter of 4 was changed to 1.
Model M033 is the alternative model as described in the Supplementary Methods.
Model M034 is the default model (M019) with an alternative, non-linear feedback from MDM2 to p53.